NiceWindow.h: fix a compile warning about int-vs-NSInteger
Since [-setLevel:] is from NSWindow anyway, I'm not sure if we even need to bother declaring it?

+++ src/app/NiceWindow.h
@@ -104,7 +104,7 @@
 -(void)makeFullScreen;
 -(void)makeNormalScreen;
 -(BOOL)isFullScreen;
--(void)setLevel:(int)windowLevel;
+-(void)setLevel:(NSInteger)windowLevel;
 -(void)resizeWithSize:(NSSize)aSize animate:(BOOL)animate;
 -(NSRect)calcResizeSize:(NSSize)aSize;
 -(void)resize:(float)amount animate:(BOOL)animate;
